fix detection of gio/gdesktopappinfo.h
Header file gio/gdesktopappinfo.h is currently never found
by Meson because it's inside gio-unix-2.0 directory.

This bug was introduced by commit 13b94dab10e which
fixed the build for MacOS but caused the header file
to not be detected by Linux anymore.

One possible fix would be to replace 'gio/gdesktopappinfo.h'
with 'gio-unix-2.0/gio/gdesktopappinfo.h', but as it's a
unix only header we move the check to an appropiate place
where other unix specific handling is done.