label: Fix measuring min wfh with wrap+ellipsize+lines
GtkLabel's layout is influenced by the available horizontal space
(allocated width), which affects both wrapping and ellipsization, but
never the available vertical space (allocated height); instead, GtkLabel
requests, as its minimum size in the vertical orientation, the exact
height that it needs and is going to use for the layout.

For a label that has all of wrap, ellipsize, and lines set, this means
that depending on the label text and the available width, the label may
be wrapped into up to the specified number of lines. If there's still
not enough space, the label will begin to ellipsize. Crucially, it won't
start ellipsizing before reaching the specified number of lines. For
example, if a label has lines = 5, it can take up anywhere between 1 to
5 lines, unellipsized, or 5 lines, ellipsized; but e.g. 2 lines,
ellipsized is not a possible layout.

This has to be taken into account when measuring the label. If we're
measuring width-for-height and the provided for_size (height) fits 2
lines, we must request enough width for the whole text to fit (into 2
lines), without any ellipsization happening. The existing logic instead
assumed that when ellipsization is enabled, minimum width could always
be set low. Change that to actually binary-search a layout width-for-
height, just like it's done in the other branch.

This fixes another case of GtkLabel reporting inconsistent width-for-
height and height-for-width measurements, which may break logic in the
rest of the layout machinery.

Signed-off-by: Sergey Bugaev <bugaevc@gmail.com>